feat: add EventManager singleton; can drag items (but not place)
The EventManager singleton is for decoupled message sending. Dragging
connections between GameObjects in the scene is not possible for spawned
prefabs, and I'd prefer not to manually send each event to the correct
singleton/components from the sender. These events lets one event source
invoke the event, and every other script that cares to listen if it
wants to.